Update Block-Email to better detect spoofed senders
Instead of just using SPF, now matches your domain to the senders domain, checks DKIM, DMARC and O365 CompAuth. Also supports the Sender EML field and attempts to use it if it detects the From field has the internal domain. The original and new email address is output to the log